refactor: add slash handler to API server and move vulnerability model to model package
