+++ api/server.go
@@ -10,6 +10,7 @@ import (
 
 func (s *APIServer) Run() {
 	router := mux.NewRouter()
+	router.HandleFunc("/", config.LogHandler(makeHTTPHandleFunc(SlashHandler))).Methods("GET")
 
 	router.HandleFunc("/api/v1/health", config.LogHandler(makeHTTPHandleFunc(HealthHandler))).Methods("GET")
 	router.HandleFunc("/api/v1/vulnerability/cve", config.LogHandler(makeHTTPHandleFunc(ReadVulnerabilityByCve))).Methods("GET")
@@ -19,6 +20,10 @@ func (s *APIServer) Run() {
 	if err != nil {
 		return
 	}
+}
+
+func SlashHandler(w http.ResponseWriter, r *http.Request) error {
+	return WriteJson(w, http.StatusOK, API{message: "Welcome to Snowden API"})
 }
 
 func NewApiServer(port string) *APIServer {
@@ -53,4 +58,8 @@ type apiFunc func(w http.ResponseWriter, r *http.Request) error
 
 type ApiError struct {
 	Error string `json:"error"`
+}
+
+type API struct {
+	message string
 }
