+++ client/nvd-client.go
@@ -5,10 +5,11 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"snowden/client/model"
 	"snowden/config"
 )
 
-func GetVulnerabilityByCveId(cveId string) (Vulnerability, error) {
+func GetVulnerabilityByCveId(cveId string) (model.Vulnerability, error) {
 	var vulnerability CveVulnerability
 	nvdUrl := config.GetEnv("NVD_URL")
 
@@ -18,7 +19,7 @@ func GetVulnerabilityByCveId(cveId string) (Vulnerability, error) {
 	resp, err := http.Get(url)
 
 	if err != nil {
-		return Vulnerability{}, err
+		return model.Vulnerability{}, err
 	}
 
 	defer func(Body io.ReadCloser) {
@@ -27,12 +28,12 @@ func GetVulnerabilityByCveId(cveId string) (Vulnerability, error) {
 
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		return Vulnerability{}, err
+		return model.Vulnerability{}, err
 	}
 
 	err = json.Unmarshal(body, &vulnerability)
 	if err != nil {
-		return Vulnerability{}, err
+		return model.Vulnerability{}, err
 	}
 
 	vulnModel := modelVulnerability(vulnerability)
@@ -68,7 +69,7 @@ type CveVulnerability struct {
 	} `json:"vulnerabilities"`
 }
 
-func modelVulnerability(vuln CveVulnerability) Vulnerability {
+func modelVulnerability(vuln CveVulnerability) model.Vulnerability {
 	vulnerability := vuln.Vulnerabilities[0]
 
 	return vulnerability
